Remove duplicated code.
